+++ examples/next-prisma-starter/src/utils/trpc.ts
@@ -111,15 +111,18 @@ export const trpc = setupTRPC<AppRouter, SSRContext>({
  * This is a helper method to infer the output of a query resolver
  * @example type HelloOutput = inferQueryOutput<'hello'>
  */
-export type inferQueryOutput<TRouteKey extends keyof AppRouter['queries']> =
-  inferProcedureOutput<AppRouter['queries'][TRouteKey]>;
+export type inferQueryOutput<
+  TRouteKey extends keyof AppRouter['_def']['queries'],
+> = inferProcedureOutput<AppRouter['_def']['queries'][TRouteKey]>;
 
-export type inferQueryInput<TRouteKey extends keyof AppRouter['queries']> =
-  inferProcedureInput<AppRouter['queries'][TRouteKey]>;
+export type inferQueryInput<
+  TRouteKey extends keyof AppRouter['_def']['queries'],
+> = inferProcedureInput<AppRouter['_def']['queries'][TRouteKey]>;
 
 export type inferMutationOutput<
-  TRouteKey extends keyof AppRouter['mutations'],
-> = inferProcedureOutput<AppRouter['mutations'][TRouteKey]>;
+  TRouteKey extends keyof AppRouter['_def']['mutations'],
+> = inferProcedureOutput<AppRouter['_def']['mutations'][TRouteKey]>;
 
-export type inferMutationInput<TRouteKey extends keyof AppRouter['mutations']> =
-  inferProcedureInput<AppRouter['mutations'][TRouteKey]>;
+export type inferMutationInput<
+  TRouteKey extends keyof AppRouter['_def']['mutations'],
+> = inferProcedureInput<AppRouter['_def']['mutations'][TRouteKey]>;
