+++ packages/client/src/index.ts
@@ -1,4 +1,5 @@
 export * from './createTRPCClient';
+export * from './createTRPCClientProxy';
 export * from './getFetch';
 export * from './TRPCClientError';
 export * from './links';
